add initial working code
This commit includes the curren version of the notebook and train_model.py
file which allows you to run a hyperparameter search and deploy a model.

It's not quite complete yet as the notebook needs to be tidied up and I want
to update the code to be able to run batch inference, but I want to make
sure that I have a backup of the code in case something goes wrong.